% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intro_for_docs.R
\name{A Brief Introduction}
\alias{A Brief Introduction}
\title{Welcome to \code{tidycells} package}
\description{
This is a quick guide on \emph{how to utilize documentation} for \code{tidycells}.
}
\section{TL;DR }{


\if{html}{If you are not sure about what package functions actually do or how they work together,
please start with \href{../doc/tidycells-intro.html}{\code{vignette("tidycells-intro")}}.}
\if{latex}{If you are not sure about what package functions actually do or how they work together,
please start with \code{vignette("tidycells-intro")}.}
}

\section{Introduction }{


\if{latex}{
\figure{logo.png}{options: width=1in}
}
\code{tidycells} hosts various functions for ...
}

