% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-etc.R
\name{df-class}
\alias{df-class}
\title{minor internal classes}
\description{
The \code{df} class is based on data.frame,
}
\section{Properties of \code{Table_Field_Container}}{


Objects of class \code{Table_Field_Container} have following named nodes:
\itemize{
\item \code{cells} : Contains information about \code{cell_group_type} in terms of (data, minor and major attributes).(a tibble)
\item \code{sections} : Contains boundaries of each data block. (a tibble)
\item \code{details} : a list containing further information
\item \code{cell_df} : The original \code{\link[=cell_df-class]{cell_df}} which is passed for processing
}
}

\section{Applicable methods on \code{Table_Field_Container}}{

\itemize{
\item \code{print}: Prints identifier of the class and the number of blocks (and potential issues if any).
\item Apart from these all methods available to data.frame or tibble.
}
}

\keyword{internal}
