% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directions.R
\name{get_direction_metric}
\alias{get_direction_metric}
\title{get direction metric}
\usage{
get_direction_metric(d1, a1, direction)
}
\arguments{
\item{d1}{part of d_part with data_gid}

\item{a1}{part of d_part with attr_gid}

\item{direction}{direction name
should be one of \code{\link[=get_valid_direction_names]{get_valid_direction_names}}}
}
\value{
a scaled fraction denoting coverage (1 means full coverage) for the supplied direction.
}
\description{
get direction metric
}
\details{
Used internally by \code{\link[=get_direction]{get_direction}} function
}
\keyword{internal}
