% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directions.R
\name{get_valid_direction_names}
\alias{get_valid_direction_names}
\title{get direction names compatible to tidycells}
\usage{
get_valid_direction_names(header_binder = getOption("tidycells.header_binder"))
}
\value{
directions as used in \code{\link[=bind_header]{bind_header}} applicant (\emph{like unpivotr package}) with directional grouping
}
\description{
get direction names compatible to tidycells
}
\details{
Kept for compatibility. Used internally by get_direction function.
}
\keyword{internal}
