% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LCS}
\alias{LCS}
\title{Longest Common Substring using dynamic programming}
\usage{
LCS(ss)
}
\arguments{
\item{ss}{a string vector.}
}
\description{
Longest Common Substring using dynamic programming
}
\examples{
LCS(c("hi","hello and hi","only hi"))
}
\keyword{internal}
