% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Table_Field.R
\name{Table_Field-class}
\alias{Table_Field-class}
\title{\code{Table_Field} class}
\description{
The \code{Table_Field} class is based on data.frame,
created in order to store analysis of cell level information (\code{\link[=cell_df-class]{cell_df}}).
}
\section{Properties of \code{Table_Field}}{


Objects of class \code{Table_Field} have following named nodes:
\itemize{
\item \code{cells} : Contains information about \code{cell_group_type} in terms of (data, minor and major attributes).(a tibble)
\item \code{sections} : Contains boundaries of each data block. (a tibble)
\item \code{details} : a list containing further information
\item \code{cell_df} : The original \code{\link[=cell_df-class]{cell_df}} which is passed for processing
}
}

\section{Applicable methods on \code{Table_Field}}{

\itemize{
\item \code{print}: Prints identifier of the class and the number of blocks (and potential issues if any).
\item Apart from these all methods available to data.frame or tibble.
}
}

\keyword{internal}
