% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycells-package.R
\docType{package}
\name{tidycells-package}
\alias{tidycells}
\alias{tidycells-package}
\title{tidycells: Read Tabular Data from Diverse Sources and Easily Make
    Them Tidy}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Provides utilities to read, cells from complex
    tabular data and heuristic detection based 'structural assignment' of
    those cells to a columnar or tidy format. Read functionality has the
    ability to read structured, partially structured or unstructured
    tabular data from various types of documents. The 'structural
    assignment' functionality has both supervised and unsupervised way of
    assigning cells data to columnar/tidy format. Multiple disconnected
    blocks of tables in a single sheet are also handled appropriately.
    These tools are suitable for unattended conversation of messy tables
    into a consumable format(usable for further analysis and data
    wrangling).
}
\section{Get Started}{

To get started check out \href{../doc/tidycells-intro.html}{\code{vignette("tidycells-intro")}}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://r-rudra.github.io/tidycells/}
  \item \url{https://github.com/r-rudra/tidycells}
  \item Report bugs at \url{https://github.com/r-rudra/tidycells/issues}
}

}
\author{
\strong{Maintainer}: Indranil Gayen \email{nil.gayen@gmail.com} (0000-0003-0197-1944)

}
\keyword{internal}
